requirements.txt
# LangChain Core
langchain
langchain-core
langchain-text-splitters



# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-genai



# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Web Interface
streamlit

# Structured Output
pydantic
typing-extensions



# Document Loaders
langchain-community

# PDF and HTML Parsing
pypdf
beautifulsoup4
requests



# Vector Store Backends
chromadb
faiss-cpu
wikipedia
ipykernel

# YouTube Transcript Support
youtube-transcript-api

# Web Search
duckduckgo-search



langchain-groq 
langchain_tavily

langgraph

langsmith
langgraph-cli[inmem]

langchain-mcp-adapters
mcp
langgraph-checkpoint-sqlite 

ddgs